+++ packages/article-byline/__tests__/article-byline.native.test.js
@@ -1,6 +1,0 @@
-/* eslint-env jest */
-
-import tests from "./test-helper";
-import ArticleByline from "../article-byline";
-
-describe("ArticleByline Native", tests(ArticleByline));