+++ packages/link/link.js
@@ -1,25 +1,12 @@
 import React from "react";
-import { Text, Linking } from "react-native";
+import { TouchableOpacity } from "react-native";
+import propTypes from "./link.proptypes";
 
-import { defaultProps, propTypes } from "./link.proptypes";
-
-const styles = {
-  color: "#069",
-  textDecorationLine: "underline"
-};
-
-export default function Link(props) {
-  const { children, style, url } = props;
-
-  const { onPress = () => Linking.openURL(url) } = props;
-
-  return (
-    <Text {...props} onPress={onPress} style={[styles, style]}>
-      {children}
-    </Text>
-  );
-}
-
-Link.defaultProps = defaultProps;
+const Link = ({ onPress, children }) => (
+  <TouchableOpacity onPress={onPress}>{children}</TouchableOpacity>
+);
 
 Link.propTypes = propTypes;
+
+export default Link;
+export { default as TextLink } from "./text-link";