+++ packages/article-byline/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.3.3"></a>
+## [0.3.3](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.3.1...@times-components/article-byline@0.3.3) (2017-09-20)
+
+
+
+
+<a name="0.3.2"></a>
+## [0.3.2](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.3.1...@times-components/article-byline@0.3.2) (2017-09-20)
+
+
+
+
+<a name="0.3.1"></a>
+## [0.3.1](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.3.0...@times-components/article-byline@0.3.1) (2017-09-19)
+
+
+
+
 <a name="0.3.0"></a>
 # [0.3.0](https://github.com/newsuk/times-components/compare/@times-components/article-byline@0.2.0...@times-components/article-byline@0.3.0) (2017-09-08)