+++ packages/article-label/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.1.1"></a>
+## [0.1.1](https://github.com/newsuk/times-components/compare/@times-components/article-label@0.1.0...@times-components/article-label@0.1.1) (2017-09-19)
+
+
+
+
 <a name="0.1.0"></a>
 # [0.1.0](https://github.com/newsuk/times-components/compare/@times-components/article-label@0.0.6...@times-components/article-label@0.1.0) (2017-09-08)